Feat: #111 - 테스트 뷰에 확대,축소,위치,회전 제스처 추가
