+++ 2024-MacC-M4-6princess/2024-MacC-M4-6princess/Source/View/ContentView.swift
@@ -13,9 +13,9 @@ struct ContentView: View {
     
     var body: some View {
         
-        CameraView()
-            .environmentObject(imageModel)
-//        IEDevelopView()
+//        CameraView()
+//            .environmentObject(imageModel)
+        IEDevelopView()
         //        CMView()
         //        IEProgressView()
     }
